Set ogr2ogr autodetect size limit to the whole file
Prior to this change, only the first 100,000 bytes of the input
file were scanned to detect field types. This led to incorrect
behavior when a dummy row was added at the end to force a numeric
field to treated as text.

This change removes the scan limit to resolve the problem at the
expense of potentially slowing down large imports.